extract hud. I'm thinking of stripping down DrawingTrait
Right now the DrawingTrait is a mix of functions that need to
be implemented by the graphics backend (macroquad) and higher-level
functions (that use the backend ones).

The higher level ones don't need to be in the DrawingTrait, as they
will accumulate over time, and I should split them into modules. As
far as I know, I can't split the method list of a struct (or trait).