handle the profiling toggle properly
I want to keep the world/screen toplevel abstraction, but that
makes it weird how the world has to handle game_state properties
like "profile", because the game state doesn't control the fluids.
World does, so world has to update both fluids and game state
when the "profile" property changes.

The alternatives are either:
- not store profiling in fluids, but
  it doesn't really fix the issue because the world still has to pass
  that to fluids from game state every frame.
- invert control so that game state contains world. ugh, the world/screen
  abstraction disappears.