fix buttons not receing click events properly
the issue was that macroquad returns that a button was clicked
when the mouse button is first pressed, and egui counts the click
only when releasing the mouse button.

I made my egui drawer to return clicked on first pressed, and had
to add some state to avoid returning click events for all the
subsequent events while the mouse button is pressed :(

I discarded the alternative of adding state to the macroquad drawer
to make it click on release because I'm not sure it's doable with
the mechanics of clicking and dragging to select world cells.