pass Box<Option<SceneState>> to frame
this panics when using macroquad context.
I assume that the library compiles its own
copy of the global variables of macroquad,
same as the main binary, and only the objects
in main get initialized.

I think the solution to make this hot-reload
bussiness work is to never use macroquad in the
lib. However, I can't split the code in
workspaces to avoid even having the macroquad dependency
in the lib, because I use extensively some types like
Vec2 and Color.