make the task queue a task stack instead of a sorted blocking queue
This way, if there's some blocked task A that needs you to create
some other thing B first, it will not require you to click "do now"
after creating B. It will just try later tasks in the queue and
come back to A later.

also, this will keep the order when rendering the task queue GUI,
and the "do now" is useless now.